requirements.txt
asgiref==3.8.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
cffi==1.16.0
channels==4.0.0
channels-redis==4.2.0
constantly==23.10.4
daphne==4.1.0
Django==5.0.3
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-api-key==2.3.0
djsm>=0.1.51
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
msgpack==1.0.8
psycopg2==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.21
pyOpenSSL==24.1.0
python-dotenv==1.0.1
PyYAML==6.0.1
redis==5.0.3
rsa==4.9
service-identity==24.1.0
setuptools==69.2.0
simple_file_handler==0.0.1
six==1.16.0
sqlparse==0.4.4
toml==0.10.2
Twisted==24.3.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.10.0
tzdata==2024.1
zope.interface==6.2
